Split resolution.ts into concern-based modules
resolution.ts was 795 lines (2× the 400 source-file cap). Extract three
cohesive concerns into sibling modules, leaving resolveNightActions as the
orchestrator:

- attack-map.ts — base attack/protection map construction and the interim
  attacked-player query (collectBaseAttacksAndProtections, buildKilledEvents,
  applyPriestWards, applyArsonistIgnite, removeFromMapSet,
  getInterimAttackedPlayerIds)
- monarch-resolution.ts — Monarch knighted-protection logic
- veteran-resolution.ts — Veteran counter-kill collection (wolf repel + visitor)

resolution.ts drops to 365 lines. No behavior change — the orchestrator's
control flow is unchanged; the only inline-to-function extraction is the
Veteran block, now collectVeteranCounterkills. Public API is preserved:
getInterimAttackedPlayerIds is re-exported from the utils barrel, and the one
direct importer (doctor-priest.spec.ts) now imports it from ./attack-map.

648 tests pass (resolution, display, werewolf actions and services); tsc and
lint clean.

Closes #652
