+++ src/lib/game/modes/werewolf/utils/resolution.ts
@@ -1,241 +1,18 @@
 import type { PlayerRoleAssignment } from "@/lib/types";
-import { Team } from "@/lib/types";
 
-import { getWerewolfRole, WerewolfRole } from "../roles";
-import type {
-  AnyNightAction,
-  NightAction,
-  NightResolutionEvent,
-} from "../types";
+import { WerewolfRole } from "../roles";
+import type { AnyNightAction, NightResolutionEvent } from "../types";
 import {
-  isTeamNightAction,
-  TargetCategory,
-  VeteranCounterkillSource,
-} from "../types";
-import { baseGroupPhaseKey, isGroupPhaseKey, isRoleActive } from "./phase-keys";
-import { getGroupPhasePlayerIds } from "./targeting";
+  applyArsonistIgnite,
+  applyPriestWards,
+  buildKilledEvents,
+  collectBaseAttacksAndProtections,
+} from "./attack-map";
+import { applyMonarchProtection } from "./monarch-resolution";
+import { collectVeteranCounterkills } from "./veteran-resolution";
 
 export const SMITE_PHASE_KEY = "__narrator_smite__";
 export const OLD_MAN_TIMER_KEY = "__old_man_timer__";
-
-function allWerewolvesAreDead(
-  roleAssignments: PlayerRoleAssignment[],
-  deadPlayerIds: string[],
-): boolean {
-  return roleAssignments
-    .filter((a) => {
-      const role = getWerewolfRole(a.roleDefinitionId);
-      return role?.isWerewolf === true;
-    })
-    .every((a) => deadPlayerIds.includes(a.playerId));
-}
-
-function chupacabraAttackApplies(
-  targetPlayerId: string,
-  roleAssignments: PlayerRoleAssignment[],
-  deadPlayerIds: string[],
-): boolean {
-  const targetAssignment = roleAssignments.find(
-    (a) => a.playerId === targetPlayerId,
-  );
-  const targetRole = targetAssignment
-    ? getWerewolfRole(targetAssignment.roleDefinitionId)
-    : undefined;
-  return (
-    targetRole?.isWerewolf === true ||
-    allWerewolvesAreDead(roleAssignments, deadPlayerIds)
-  );
-}
-
-/** Removes all occurrences of `value` from `map[key]`, deleting the key if it becomes empty. */
-function removeFromMapSet(
-  map: Map<string, string[]>,
-  key: string,
-  value: string,
-): void {
-  const existing = map.get(key) ?? [];
-  const filtered = existing.filter((v) => v !== value);
-  if (filtered.length === 0) {
-    map.delete(key);
-  } else {
-    map.set(key, filtered);
-  }
-}
-
-/**
- * Collects attacks and protections from all non-Witch, non-Spellcaster actions.
- * Returns the base attack/protect maps used by both full resolution and the
- * interim attacked-player query for the Witch.
- */
-function collectBaseAttacksAndProtections(
-  nightActions: Record<string, AnyNightAction>,
-  roleAssignments: PlayerRoleAssignment[],
-  deadPlayerIds: string[],
-  mirrorcasterCharged?: boolean,
-  mercenaryCharged?: boolean,
-): {
-  attacks: Map<string, string[]>;
-  protections: Map<string, string[]>;
-} {
-  const attacks = new Map<string, string[]>();
-  const protections = new Map<string, string[]>();
-
-  for (const [phaseKey, action] of Object.entries(nightActions)) {
-    // Witch and Spellcaster are handled separately below.
-    // Priest protection is handled via priestWards, not the generic Protect pipeline.
-    // Altruist intercept is applied after this loop.
-    // Swapper swap is applied after Altruist intercept.
-    if (
-      isRoleActive(phaseKey, [
-        WerewolfRole.Witch,
-        WerewolfRole.Spellcaster,
-        WerewolfRole.Priest,
-        WerewolfRole.Altruist,
-        WerewolfRole.Swapper,
-      ])
-    )
-      continue;
-
-    if (isGroupPhaseKey(phaseKey)) {
-      const groupAction = action as { suggestedTargetId?: string };
-      if (!groupAction.suggestedTargetId) continue;
-      const tid = groupAction.suggestedTargetId;
-      attacks.set(tid, [...(attacks.get(tid) ?? []), phaseKey]);
-      continue;
-    }
-
-    const soloAction = action as { targetPlayerId?: string };
-    if (!soloAction.targetPlayerId) continue;
-    const tid = soloAction.targetPlayerId;
-
-    const role = getWerewolfRole(phaseKey);
-    if (!role) continue;
-
-    if (role.targetCategory === TargetCategory.Protect) {
-      protections.set(tid, [...(protections.get(tid) ?? []), phaseKey]);
-      continue;
-    }
-
-    if (role.targetCategory === TargetCategory.Attack) {
-      if (
-        isRoleActive(phaseKey, WerewolfRole.Chupacabra) &&
-        !chupacabraAttackApplies(tid, roleAssignments, deadPlayerIds)
-      ) {
-        continue;
-      }
-      attacks.set(tid, [...(attacks.get(tid) ?? []), phaseKey]);
-      continue;
-    }
-
-    // Mirrorcaster: acts as Protect when uncharged, Attack when charged.
-    if (isRoleActive(phaseKey, WerewolfRole.Mirrorcaster)) {
-      if (mirrorcasterCharged) {
-        attacks.set(tid, [...(attacks.get(tid) ?? []), phaseKey]);
-      } else {
-        protections.set(tid, [...(protections.get(tid) ?? []), phaseKey]);
-      }
-    }
-
-    // Mercenary: acts as Protect when uncharged (no combat effect when charged — bribe only).
-    if (isRoleActive(phaseKey, WerewolfRole.Mercenary)) {
-      if (!mercenaryCharged) {
-        protections.set(tid, [...(protections.get(tid) ?? []), phaseKey]);
-      }
-    }
-  }
-
-  return { attacks, protections };
-}
-
-function buildKilledEvents(
-  attacks: Map<string, string[]>,
-  protections: Map<string, string[]>,
-): NightResolutionEvent[] {
-  return Array.from(attacks.entries()).map(([targetPlayerId, attackedBy]) => {
-    const protectedBy = protections.get(targetPlayerId) ?? [];
-    return {
-      type: "killed" as const,
-      targetPlayerId,
-      attackedBy,
-      protectedBy,
-      died: protectedBy.length === 0,
-    };
-  });
-}
-
-/** Adds priest ward protections to the protections map for any warded player under attack. */
-function applyPriestWards(
-  attacks: Map<string, string[]>,
-  protections: Map<string, string[]>,
-  priestWards: Record<string, string>,
-): void {
-  for (const wardedPlayerId of Object.keys(priestWards)) {
-    if (attacks.has(wardedPlayerId)) {
-      protections.set(wardedPlayerId, [
-        ...(protections.get(wardedPlayerId) ?? []),
-        WerewolfRole.Priest,
-      ]);
-    }
-  }
-}
-
-/** Adds Arsonist ignite attacks to the attacks map when the Arsonist self-targeted. */
-function applyArsonistIgnite(
-  attacks: Map<string, string[]>,
-  nightActions: Record<string, AnyNightAction>,
-  roleAssignments: PlayerRoleAssignment[],
-  arsonistDousedPlayerIds: string[] | undefined,
-): void {
-  if (!arsonistDousedPlayerIds?.length) return;
-  const arsonistAction = nightActions[WerewolfRole.Arsonist] as
-    | { targetPlayerId?: string }
-    | undefined;
-  const arsonistPlayerId = roleAssignments.find(
-    (a) => a.roleDefinitionId === (WerewolfRole.Arsonist as string),
-  )?.playerId;
-  if (
-    arsonistAction?.targetPlayerId &&
-    arsonistAction.targetPlayerId === arsonistPlayerId
-  ) {
-    for (const dousedId of arsonistDousedPlayerIds) {
-      attacks.set(dousedId, [
-        ...(attacks.get(dousedId) ?? []),
-        WerewolfRole.Arsonist,
-      ]);
-    }
-  }
-}
-
-/**
- * Returns player IDs who are currently attacked but not yet protected,
- * excluding the Witch's own action. Used to show the Witch their available
- * targets before they act. Also considers priest wards and Arsonist ignite.
- */
-export function getInterimAttackedPlayerIds(
-  nightActions: Record<string, AnyNightAction>,
-  roleAssignments: PlayerRoleAssignment[],
-  deadPlayerIds: string[],
-  priestWards?: Record<string, string>,
-  mirrorcasterCharged?: boolean,
-  mercenaryCharged?: boolean,
-  arsonistDousedPlayerIds?: string[],
-): string[] {
-  const { attacks, protections } = collectBaseAttacksAndProtections(
-    nightActions,
-    roleAssignments,
-    deadPlayerIds,
-    mirrorcasterCharged,
-    mercenaryCharged,
-  );
-  applyArsonistIgnite(
-    attacks,
-    nightActions,
-    roleAssignments,
-    arsonistDousedPlayerIds,
-  );
-  if (priestWards) applyPriestWards(attacks, protections, priestWards);
-  return Array.from(attacks.keys()).filter((id) => !protections.has(id));
-}
 
 /**
  * Resolves all night actions into a flat list of outcome events.
@@ -270,93 +47,6 @@ export interface NightResolutionOptions {
     monarchPlayerId: string;
     monarchKnightedPlayerIds: string[];
   };
-}
-
-function getAttackerIds(
-  attackedBy: string[],
-  roleAssignments: PlayerRoleAssignment[],
-  deadPlayerIds: string[],
-): string[] {
-  const attackerIds = new Set<string>();
-  for (const phaseKey of attackedBy) {
-    if (isGroupPhaseKey(phaseKey)) {
-      for (const playerId of getGroupPhasePlayerIds(
-        roleAssignments,
-        phaseKey,
-        deadPlayerIds,
-      )) {
-        attackerIds.add(playerId);
-      }
-      continue;
-    }
-    const attacker = roleAssignments.find(
-      (assignment) =>
-        assignment.roleDefinitionId === phaseKey &&
-        !deadPlayerIds.includes(assignment.playerId),
-    );
-    if (attacker) attackerIds.add(attacker.playerId);
-  }
-  return [...attackerIds];
-}
-
-function applyMonarchProtection(
-  attacks: Map<string, string[]>,
-  protections: Map<string, string[]>,
-  roleAssignments: PlayerRoleAssignment[],
-  deadPlayerIds: string[],
-  monarchProtection: {
-    monarchPlayerId: string;
-    monarchKnightedPlayerIds: string[];
-  },
-): void {
-  const { monarchPlayerId, monarchKnightedPlayerIds } = monarchProtection;
-  if (!attacks.has(monarchPlayerId)) return;
-
-  const livingKnightedPlayerIds = monarchKnightedPlayerIds.filter(
-    (playerId) => !deadPlayerIds.includes(playerId),
-  );
-  if (livingKnightedPlayerIds.length === 0) return;
-
-  const attackerIds = getAttackerIds(
-    attacks.get(monarchPlayerId) ?? [],
-    roleAssignments,
-    deadPlayerIds,
-  );
-  const getRoleForPlayer = (playerId: string) => {
-    const roleDefinitionId = roleAssignments.find(
-      (assignment) => assignment.playerId === playerId,
-    )?.roleDefinitionId;
-    return roleDefinitionId ? getWerewolfRole(roleDefinitionId) : undefined;
-  };
-  const livingKnightedRoleDefs = livingKnightedPlayerIds
-    .map((playerId) => getRoleForPlayer(playerId))
-    .filter((roleDef) => roleDef !== undefined);
-  const allLivingKnightedPlayersAreBad = livingKnightedRoleDefs.every(
-    (roleDef) => roleDef.team === Team.Bad,
-  );
-  const badAttackerExists = attackerIds.some((playerId) => {
-    const attackerRole = getRoleForPlayer(playerId);
-    return attackerRole?.team === Team.Bad;
-  });
-  const onlyLivingKnightedPlayerId =
-    livingKnightedPlayerIds.length === 1
-      ? livingKnightedPlayerIds[0]
-      : undefined;
-  const attackerIsOnlyLivingKnightedPlayer =
-    onlyLivingKnightedPlayerId !== undefined &&
-    attackerIds.includes(onlyLivingKnightedPlayerId);
-
-  if (
-    (badAttackerExists && allLivingKnightedPlayersAreBad) ||
-    attackerIsOnlyLivingKnightedPlayer
-  ) {
-    return;
-  }
-
-  protections.set(monarchPlayerId, [
-    ...(protections.get(monarchPlayerId) ?? []),
-    WerewolfRole.Monarch,
-  ]);
 }
 
 export function resolveNightActions(
@@ -500,138 +190,18 @@ export function resolveNightActions(
     }
   }
 
-  // Veteran alert: if the Veteran alerted this night, resolve counter-kills.
-  // Counter-kills happen after the Altruist and Swapper so both cannot intercept them.
-  const veteranAction = nightActions[WerewolfRole.Veteran];
-  const veteranAlerted =
-    veteranAction !== undefined &&
-    !isTeamNightAction(veteranAction) &&
-    veteranAction.alerted === true;
-  // Pending counter-kill entries; events are emitted after Tough Guy so the
-  // `died` field reflects the actual outcome rather than the initial attack.
-  const pendingVeteranKills: {
-    counterkilledPlayerId: string;
-    veteranPlayerId: string;
-    source: VeteranCounterkillSource;
-  }[] = [];
-
-  if (veteranAlerted) {
-    const veteranPlayerId = roleAssignments.find(
-      (a) => a.roleDefinitionId === (WerewolfRole.Veteran as string),
-    )?.playerId;
-
-    if (veteranPlayerId) {
-      // Werewolf attack repel: if a wolf group's attack is currently on the
-      // Veteran (post-Swapper), remove it and counter-kill one alive wolf
-      // participant instead. Iterating over the attacks map (rather than
-      // nightActions.suggestedTargetId) ensures Swapper-redirected wolf
-      // attacks are caught as well as directly-targeted ones.
-      // Track already-selected victims so multiple wolf-group phases don't
-      // select the same wolf when there are bonus attack keys (e.g. Wolf Cub).
-      const selectedWolfVictimIds = new Set<string>();
-      const wolfPhaseKeysOnVeteran = (
-        attacks.get(veteranPlayerId) ?? []
-      ).filter(isGroupPhaseKey);
-      for (const phaseKey of wolfPhaseKeysOnVeteran) {
-        // Remove this wolf-group's attack entry from the Veteran.
-        removeFromMapSet(attacks, veteranPlayerId, phaseKey);
-
-        // Find the first alive participant in this wolf group not already
-        // selected as a victim this resolution pass.
-        const baseKey = baseGroupPhaseKey(phaseKey);
-        const wolfVictimId = roleAssignments.find((a) => {
-          if (deadPlayerIds.includes(a.playerId)) return false;
-          if (selectedWolfVictimIds.has(a.playerId)) return false;
-          if (a.roleDefinitionId === baseKey) return true;
-          const role = getWerewolfRole(a.roleDefinitionId);
-          return (role?.wakesWith as string | undefined) === baseKey;
-        })?.playerId;
-
-        if (wolfVictimId) {
-          selectedWolfVictimIds.add(wolfVictimId);
-          attacks.set(wolfVictimId, [
-            ...(attacks.get(wolfVictimId) ?? []),
-            WerewolfRole.Veteran,
-          ]);
-          // If the wolf victim has an active Priest ward, apply the protection
-          // now so the ward absorbs this newly-queued counter-kill attack.
-          // (applyPriestWards ran before Veteran resolution and only covers
-          // attacks that existed at that point.)
-          if (options?.priestWards?.[wolfVictimId]) {
-            protections.set(wolfVictimId, [
-              ...(protections.get(wolfVictimId) ?? []),
-              WerewolfRole.Priest,
-            ]);
-          }
-          pendingVeteranKills.push({
-            counterkilledPlayerId: wolfVictimId,
-            veteranPlayerId,
-            source: VeteranCounterkillSource.WolfRepel,
-          });
-        }
-      }
-
-      // Visitor counter-kill: only Protect- and Attack-category solo roles
-      // physically visit the Veteran; Investigation and Special roles (Seer,
-      // Exposer, Spellcaster, Mummy, Dracula, Zombie, …) use mystical powers
-      // or non-physical means and are not counter-killed.
-      // Mirrorcaster is classified as Special but effectively acts as Protect
-      // (uncharged) or Attack (charged) — treat it as a physical visitor.
-      // Mercenary is classified as Special but acts as Protect when uncharged
-      // (Protect mode) — treat it as a physical visitor in that mode.
-      for (const [phaseKey, action] of Object.entries(nightActions)) {
-        if (isGroupPhaseKey(phaseKey)) continue;
-        if (isRoleActive(phaseKey, WerewolfRole.Priest)) continue;
-        const soloAction = action as NightAction;
-        if (soloAction.targetPlayerId !== veteranPlayerId) continue;
-        const role = getWerewolfRole(phaseKey);
-        if (
-          role?.targetCategory !== TargetCategory.Protect &&
-          role?.targetCategory !== TargetCategory.Attack &&
-          !isRoleActive(phaseKey, WerewolfRole.Mirrorcaster) &&
-          !(
-            isRoleActive(phaseKey, WerewolfRole.Mercenary) &&
-            !options?.mercenaryCharged
-          )
-        )
-          continue;
-
-        const visitorPlayerId = roleAssignments.find(
-          (a) => a.roleDefinitionId === phaseKey,
-        )?.playerId;
-        if (!visitorPlayerId || deadPlayerIds.includes(visitorPlayerId))
-          continue;
-
-        // Discard any protection the visitor provided to the Veteran.
-        removeFromMapSet(protections, veteranPlayerId, phaseKey);
-
-        // Remove any attack the visitor had on the Veteran (attack roles that
-        // targeted the Veteran are repelled, not just protectors).
-        removeFromMapSet(attacks, veteranPlayerId, phaseKey);
-
-        // Queue the counter-kill attack.
-        attacks.set(visitorPlayerId, [
-          ...(attacks.get(visitorPlayerId) ?? []),
-          WerewolfRole.Veteran,
-        ]);
-        // If the visitor has an active Priest ward, apply the protection now so
-        // the ward absorbs this newly-queued counter-kill attack.
-        // (applyPriestWards ran before Veteran resolution and only covers
-        // attacks that existed at that point.)
-        if (options?.priestWards?.[visitorPlayerId]) {
-          protections.set(visitorPlayerId, [
-            ...(protections.get(visitorPlayerId) ?? []),
-            WerewolfRole.Priest,
-          ]);
-        }
-        pendingVeteranKills.push({
-          counterkilledPlayerId: visitorPlayerId,
-          veteranPlayerId,
-          source: VeteranCounterkillSource.Visitor,
-        });
-      }
-    }
-  }
+  // Veteran alert: resolve counter-kills against the final attack/protection
+  // maps. Runs after the Altruist and Swapper so neither can intercept them.
+  // Events are emitted after Tough Guy (below) so `died` reflects the outcome.
+  const pendingVeteranKills = collectVeteranCounterkills(
+    nightActions,
+    roleAssignments,
+    deadPlayerIds,
+    attacks,
+    protections,
+    options?.priestWards,
+    options?.mercenaryCharged,
+  );
 
   let combatEvents = buildKilledEvents(attacks, protections);
   for (const smitedId of smitedPlayerIds ?? []) {
